remove useless constants file
