--- src/constants/index.js
@@ -1,10 +1,0 @@
-const PRIVACY = Object.freeze({
-  public: "PUBLIC",
-  private: "PRIVATE",
-  friends: "FRIENDS",
-  mutualFriends: "FOAF",
-  server: "SERVERONLY",
-  unlisted: "UNLISTED",
-});
-
-export default PRIVACY;
